@@ -14,7 +14,6 @@ classifiers = [
   "License :: OSI Approved :: MIT License",
   "Programming Language :: Python",
   "Programming Language :: Python :: 3",
-  "Programming Language :: Python :: 3.9",
   "Programming Language :: Python :: 3.10",
   "Programming Language :: Python :: 3.11",
   "Programming Language :: Python :: 3.12",
@@ -25,10 +24,10 @@ classifiers = [
 ]
 
 [tool.poetry.dependencies]
-python = "^3.9"
+python = "^3.10"
 apispec = "^6"
 marshmallow = "^3"
-muffin = "^0"
+muffin = "*"
 
 # Optional dependencies
 pyyaml = { version = "*", optional = true }
@@ -78,7 +77,7 @@ ignore_missing_imports = true
 [tool.tox]
 legacy_tox_ini = """
 [tox]
-envlist = py310,py311,py312,py313,pypy39
+envlist = py310,py311,py312,py313,pypy310
 
 [testenv]
 deps = -e .[tests]
@@ -101,9 +100,10 @@ ignore = [
   "D",
   "DJ",
   "EM",
-  "FIX",
   "FA100",
+  "FIX",
   "N804",
+  "PLC0415",
   "PLR0912",
   "PLR2004",
   "RET",
