@@ -17,24 +17,24 @@ Api = API
 
 __all__ = (
     "API",
-    "Api",
-    "RESTHandler",
     "APIError",
-    "PWRESTHandler",
-    "PWFilter",
-    "PWFilters",
-    "PWSort",
-    "PWSorting",
-    "SARESTHandler",
-    "SAFilter",
-    "SAFilters",
-    "SASort",
-    "SASorting",
-    "MongoRESTHandler",
+    "Api",
     "MongoFilter",
     "MongoFilters",
+    "MongoRESTHandler",
     "MongoSort",
     "MongoSorting",
+    "PWFilter",
+    "PWFilters",
+    "PWRESTHandler",
+    "PWSort",
+    "PWSorting",
+    "RESTHandler",
+    "SAFilter",
+    "SAFilters",
+    "SARESTHandler",
+    "SASort",
+    "SASorting",
 )
 
 # Support Peewee ORM
